dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Document-getElementsByTagName.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Document.getElementsByTagName</title>
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-document-getelementsbytagname">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
// TODO: getElementsByTagName("*")
test(function() {
  assert_false(document.getElementsByTagName("html") instanceof NodeList, "NodeList")
  assert_true(document.getElementsByTagName("html") instanceof HTMLCollection, "HTMLCollection")
  var firstCollection = document.getElementsByTagName("html"),
      secondCollection = document.getElementsByTagName("html")
  assert_true(firstCollection !== secondCollection || firstCollection === secondCollection,
              "Caching is allowed.")
})
test(function() {
  var l = document.getElementsByTagName("nosuchtag")
  l[5] = "foopy"
  assert_equals(l[5], undefined)
  assert_equals(l.item(5), null)
}, "Shouldn't be able to set unsigned properties on a HTMLCollection")
test(function() {
  assert_equals(document.createElementNS("http://www.w3.org/1999/xhtml", "i").localName, "i") // Sanity
  var i = document.body.appendChild(document.createElementNS("http://www.w3.org/1999/xhtml", "I"))
  assert_equals(i.localName, "I")
  assert_equals(i.tagName, "I")
  assert_equals(document.getElementsByTagName("I").length, 0)
  assert_equals(document.getElementsByTagName("i").length, 0)
  assert_equals(document.body.getElementsByTagName("I").length, 0)
  assert_equals(document.body.getElementsByTagName("i").length, 0)
  document.body.removeChild(i)
})
test(function() {
  var t = document.body.appendChild(document.createElementNS("test", "te:st"))
  assert_equals(document.getElementsByTagName("st").length, 1)
  assert_equals(document.getElementsByTagName("st")[0], t)
  assert_equals(document.getElementsByTagName("te:st").length, 0)
  document.body.removeChild(t)
})
test(function() {
  var h = document.body.appendChild(document.createElementNS("http://www.w3.org/1999/xhtml", "te:st"))
  assert_equals(document.getElementsByTagName("st")[0], h)
  assert_equals(document.getElementsByTagName("TE:ST").length, 0)
  document.body.removeChild(h)
})
</script>
